Moved bidding test to a new submissions end-to-end test file.
We have been creating one end-to-end test file per route, and expect to add more tests for the submissions page.
